+++ src/commands/lowLevelCommands/list-beta-testers.ts
@@ -1,0 +1,18 @@
+'use strict';
+
+import { LowLevelCommandBuilder, IRequiredInputParameter } from '../../utils/lowLevelCommandBuilder';
+import { CommandRunner } from '../../utils/commandRunner';
+import { commandInputParameters } from '../../utils/basicCommandInputParameters';
+import { OPERATION } from '../../utils/configuration';
+
+export const listBetaTesters = async () => {
+    const requiredParameters: IRequiredInputParameter[] = [];
+    requiredParameters.push(commandInputParameters.skillId);
+
+    const command = await LowLevelCommandBuilder.buildCommand(
+        OPERATION.LOW_LEVEL.LIST_BETA_TESTERS.COMMAND,
+        OPERATION.LOW_LEVEL.LIST_BETA_TESTERS.SUB_COMMAND,
+        requiredParameters
+    );
+    CommandRunner.runCommand(command);
+};