+++ CHANGELOG.md
@@ -3,3 +3,9 @@ v1.0.0
 - Added UI for runtime selection (Python / Node.js) when creating a new skill.
 - Added Python SDK Snippets suggestions.
 - ASK toolkit is now Generally Available.
+
+## Feb 10, 2019
+v1.0.1
+- Support one high level command: ask dialog.
+- Support some low level commands in beta-test and catalog category.
+- Add a new snippet for a new AMAZON built-in intent: SelectIntent.